refactor(fs): move FileHandle class
